OfflineEventReplicator in Spring Boot (#1041)
* OfflineEventReplicator in Spring Boot
The mechanism of interpreting the .services of the external REST services was not working for the spring boot case.
In particular this causes the OfflineEventReplicator service to not work
Issue: 206569

* OfflineEventReplicator in Spring Boot
The mechanism of interpreting the .services of the external REST services was not working for the spring boot case.
In particular this causes the OfflineEventReplicator service to not work
Issue: 206569

* OfflineEventReplicator in Spring Boot
The mechanism of interpreting the .services of the external REST services was not working for the spring boot case.
In particular this causes the OfflineEventReplicator service to not work
Issue: 206569